+++ lib/jnpr/junos/device.py
@@ -1025,7 +1025,8 @@ class Device(_Connection):
     # -----------------------------------------------------------------------
 
     def __new__(cls, *args, **kwargs):
-        if kwargs.get('port') in [23, '23'] or kwargs.get('mode'):
+        if kwargs.get('port') in [23, '23'] or kwargs.get('mode') or \
+                        kwargs.get('cs_user') is not None:
             from jnpr.junos.console import Console
             instance = object.__new__(Console, *args, **kwargs)
             # Python only calls __init__() if the object returned from
@@ -1049,7 +1050,11 @@ class Device(_Connection):
                              alternative for **host**
 
         :param str host:
-            **REQUIRED** host-name or ipaddress of target device
+            **REQUIRED** host-name or ipaddress of target device, unless sock_fd is provided
+
+        :param str sock_fd:
+            **REQUIRED** file descriptor of an existing socket instead of providing a host.
+            Used for outbound ssh. 
 
         :param str user:
             *OPTIONAL* login user-name, uses $USER if not provided
@@ -1113,6 +1118,7 @@ class Device(_Connection):
         hostname = vargs[0] if len(vargs) else kvargs.get('host')
 
         self._port = kvargs.get('port', 830)
+        self._sock_fd = kvargs.get('sock_fd', None)
         self._gather_facts = kvargs.get('gather_facts', True)
         self._normalize = kvargs.get('normalize', False)
         self._auto_probe = kvargs.get('auto_probe', self.__class__.auto_probe)
@@ -1134,10 +1140,12 @@ class Device(_Connection):
             self._ssh_config = None
         else:
             # --------------------------
-            # making a remote connection
+            # making a remote connection 
+            # if hostname is None, this is an 'outbound-ssh' connection
+            # which uses the established TCP connection from sock_fd
             # --------------------------
-            if hostname is None:
-                raise ValueError("You must provide the 'host' value")
+            if hostname is None and self._sock_fd is None: 
+                raise ValueError("You must provide either 'host' or 'sock_fd' value")
             self._hostname = hostname
             # user will default to $USER
             self._auth_user = os.getenv('USER')
@@ -1242,6 +1250,7 @@ class Device(_Connection):
             self._conn = netconf_ssh.connect(
                 host=self._hostname,
                 port=self._port,
+                sock_fd=self._sock_fd,
                 username=self._auth_user,
                 password=self._auth_password,
                 hostkey_verify=False,