+++ lib/jnpr/junos/transport/tty.py
@@ -51,7 +51,7 @@ class Terminal(object):
         '(?P<passwd>assword:\s*$)',
         '(?P<badpasswd>ogin incorrect)',
         '(?P<netconf_closed><!-- session end at .*-->\s*)',
-        '(?P<shell>%|#|/$\s*$)',
+        '(?P<shell>%|#|(~\$)\s*$)',
         '(?P<cli>[^\\-"]>\s*$)',
         '(?P<option>Enter your option:\s*$)',
         '(?P<hotkey>connection: <CTRL>Z)',
@@ -78,8 +78,8 @@ class Terminal(object):
         self.hostname = self.__dict__.get('host')
         self.user = kvargs.get('user', 'root')
         self.passwd = kvargs.get('passwd', '')
-        self.c_user = kvargs.get('s_user', self.user)
-        self.c_passwd = kvargs.get('s_passwd', self.passwd)
+        self.cs_user = kvargs.get('cs_user')
+        self.cs_passwd = kvargs.get('cs_passwd')
         self.login_attempts = kvargs.get('attempts') or self.LOGIN_RETRY
         self.console_has_banner = kvargs.get('console_has_banner') or False